refactoring the availability commands into their own module
