+++ src/commands/leagueInfo.js
@@ -1,0 +1,4 @@
+//------------------------------------------------------------------------------
+// Commands related to league information
+//------------------------------------------------------------------------------
+
